third-party/mcp-odoo/requirements.txt
# Core dependencies
mcp>=1.0.0
requests>=2.31.0
PyYAML>=6.0.1

# Image processing
Pillow>=10.0.0

# Enhanced HTTP client with retry logic
urllib3>=2.0.0

# Configuration management
python-dotenv>=1.0.0

# Data validation (optional but recommended)
pydantic>=2.0.0

# Testing dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development tools (optional)
# black>=23.0.0
# flake8>=6.0.0
# mypy>=1.0.0
